Renaming LemonSlice utils file to api.
